@@ -8,24 +8,26 @@
     "@material-ui/lab": "^4.0.0-alpha.53",
     "@material-ui/pickers": "v4.0.0-alpha.7",
     "@material-ui/styles": "^4.8.2",
+    "@aws-amplify/auth": "^3.4.8",
+    "@aws-amplify/core": "3.8.0",
     "@reduxjs/toolkit": "^1.3.5",
-    "axios": "^0.20.0",
-    "notistack": "^0.9.7",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "axios": "0.21.0",
+    "history": "5.0.0",
+    "notistack": "1.0.1",
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
     "react-loader-spinner": "^3.1.14",
-    "react-paginate": "^6.4.0",
-    "react-redux": "^7.1.3",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.3",
+    "react-paginate": "6.5.0",
+    "react-redux": "7.2.2",
+    "react-router-dom": "6.0.0-beta.0",
+    "react-scripts": "4.0.0",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
     "socket.io-client": "^2.3.1",
-    "uuid": "^8.3.0"
+    "uuid": "8.3.1"
   },
   "devDependencies": {
-    "clsx": "^1.1.1",
-    "eslint": "^6.6.0"
+    "clsx": "^1.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
